refactor: export Thunderstore mod card to dedicated component file
